+++ packages/firebase_ai/firebase_ai/lib/firebase_ai.dart
@@ -116,4 +116,5 @@ export 'src/tool.dart'
         Tool,
         ToolConfig,
         GoogleSearch,
-        CodeExecution;
+        CodeExecution,
+        UrlContext;