+++ packages/firebase_ai/firebase_ai/lib/src/base_model.dart
@@ -29,6 +29,7 @@ import 'client.dart';
 import 'content.dart';
 import 'developer/api.dart';
 import 'error.dart';
+import 'firebaseai_version.dart';
 import 'imagen/imagen_api.dart';
 import 'imagen/imagen_content.dart';
 import 'imagen/imagen_edit.dart';
@@ -36,7 +37,6 @@ import 'imagen/imagen_reference.dart';
 import 'live_api.dart';
 import 'live_session.dart';
 import 'tool.dart';
-import 'vertex_version.dart';
 
 part 'generative_model.dart';
 part 'imagen/imagen_model.dart';