+++ packages/firebase_data_connect/firebase_data_connect/example/lib/main.dart
@@ -41,14 +41,14 @@ void main() async {
       // 1. Debug provider
       // 2. Safety Net provider
       // 3. Play Integrity provider
-      androidProvider: AndroidProvider.debug,
+      providerAndroid: const AndroidDebugProvider(),
       // Default provider for iOS/macOS is the Device Check provider. You can use the "AppleProvider" enum to choose
       // your preferred provider. Choose from:
       // 1. Debug provider
       // 2. Device Check provider
       // 3. App Attest provider
       // 4. App Attest provider with fallback to Device Check provider (App Attest provider is only available on iOS 14.0+, macOS 14.0+)
-      appleProvider: AppleProvider.appAttest,
+      providerApple: const AppleAppAttestProvider(),
     );
   }
   if (configureEmulator) {