+++ packages/firebase_crashlytics/firebase_crashlytics_platform_interface/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 3.8.14
+
+ - Update a dependency to the latest release.
+
+## 3.8.13
+
+ - Update a dependency to the latest release.
+
 ## 3.8.12
 
  - Update a dependency to the latest release.