+++ packages/firebase_ml_model_downloader/firebase_ml_model_downloader_platform_interface/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.1.5+14
+
+ - Update a dependency to the latest release.
+
+## 0.1.5+13
+
+ - Update a dependency to the latest release.
+
 ## 0.1.5+12
 
  - Update a dependency to the latest release.