+++ packages/firebase_storage/firebase_storage_web/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 3.10.21
+
+ - Update a dependency to the latest release.
+
+## 3.10.20
+
+ - Update a dependency to the latest release.
+
 ## 3.10.19
 
  - Update a dependency to the latest release.